Change behaviour and number of handlers
There is now three log handlers:

* on console
* on a given file
* on a file in a directory based on the command

The console handler is always set (but can be disabled with 'none'
loglevel). The others are only initialized if some options/parameters
are set.

It is also now possible to pass some parameters to the init fonction for
configuring theses behaviours but CLI arguments have precedence.